% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_TMTI_inf}
\alias{calc_TMTI_inf}
\title{Leading NA}
\usage{
calc_TMTI_inf(pvals)
}
\arguments{
\item{pvals}{A NumericVector}
}
\description{
Returns the TMTI_infinity statistic from a pre-sorted vector of p-values
}
