% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_truncatedTMTI_inf}
\alias{calc_truncatedTMTI_inf}
\title{Leading NA}
\usage{
calc_truncatedTMTI_inf(pvals, m_full)
}
\arguments{
\item{pvals}{A NumericVector containing the truncated sorted p-values. It
is important that this vector: 1) contains only the truncated p-values (i.e,
those that fall below the truncation point) and 2) is sorted.}

\item{m_full}{The total (i.e., non-truncated) number of p-values.}
}
\description{
Returns the truncated TMTI_infinity statistic from pre-sorted,
pre-truncated p-values.
}
