% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FullCTP_C}
\alias{FullCTP_C}
\title{Leading NA}
\usage{
FullCTP_C(LocalTest, f, pvals)
}
\arguments{
\item{LocalTest}{A function that returns a double in (0, 1).}

\item{f}{A function that iterates LocalTest over the relevant test tree.
In practice, this is called as TestSet_C.}

\item{pvals}{A vector of p-values.}
}
\description{
Tests a user-specified subset in a CTP, using a user-supplied local test
}
