% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDown_TMTI.R
\name{TopDown_TMTI}
\alias{TopDown_TMTI}
\title{TopDown TMTI algorithm for estimating a 1-alpha confidence set for the number
of false hypotheses among a set.}
\usage{
TopDown_TMTI(
  pvals,
  subset = NULL,
  alpha = 0.05,
  gammaList = NULL,
  verbose = TRUE,
  tau = NULL,
  K = NULL,
  is.sorted = FALSE,
  mc.cores = 1L,
  chunksize = 4 * mc.cores,
  ...
)
}
\arguments{
\item{pvals}{A vector of p-values.}

\item{subset}{Numeric vector specifying a subset a p-values to estimate a
confidence set for the number of false hypotheses for. Defaults to NULL
corresponding to estimating a confidence set for the number of false
hypotheses in the entire set.}

\item{alpha}{Level in [0,1] at which to generate confidence set. Defaults
to 0.05.}

\item{gammaList}{List of pre-specified gamma functions. If NULL, the
functions will be approximated by bootstrap assuming independence. Defaults
to NULL.}

\item{verbose}{Logical, indicating whether or not to write out the progress.
Defaults to TRUE.}

\item{tau}{Numerical (in (0,1)); threshold to use in tTMTI. If set to NULL,
then either TMTI (default) or rtTMTI is used.}

\item{K}{Integer; Number of smallest p-values to use in rtTMTI. If se to NULL,
then either TMTI (default) or tTMTI is used.}

\item{is.sorted}{Logical, indicating whether the supplied p-values are already
is.sorted. Defaults to FALSE.}

\item{mc.cores}{Number of cores to parallelize onto.}

\item{chunksize}{Integer indicating the size of chunks to parallelize. E.g.,
if setting chunksize = mc.cores, each time a parallel computation is set up,
each worker will perform only a single task. If mc.cores > chunksize, some
threads will be inactive.}

\item{...}{Additional parameters.}
}
\value{
A 1-alpha lower bound for the number of false hypotheses among the
set of supplied p-values
}
\description{
TopDown TMTI algorithm for estimating a 1-alpha confidence set for the number
of false hypotheses among a set.
}
\examples{
## Simulate some p-values
## The first 10 are from false hypotheses, the next 10 are from true
pvals = c(
  rbeta(10, 1, 20), ## Mean value of .05
  runif(10)
)
TopDown_TMTI(pvals)

}
