% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_TMTI.R
\name{adjust_TMTI}
\alias{adjust_TMTI}
\title{Adjust all marginally significant using a Closed Testing Procedeure and the
TMTI family of tests.}
\usage{
adjust_TMTI(
  pvals,
  alpha = 0.05,
  B = 1000,
  gammaList = NULL,
  tau = NULL,
  K = NULL,
  is.sorted = FALSE,
  earlyStop = FALSE,
  verbose = FALSE,
  mc.cores = 1L,
  chunksize = 4 * mc.cores,
  ...
)
}
\arguments{
\item{pvals}{vector of pvalues}

\item{alpha}{signicance level. Defaults to 0.05}

\item{B}{Number of bootstrap replications. Only relevant if length(pvals) > 100
and no gammaList is supplied}

\item{gammaList}{A list of functions. These functions should be the CDFs of
the chosen TMTI test for different m.}

\item{tau}{Number between 0 and 1 or NULL, describing the truncation level.}

\item{K}{Integer between >1 and m describing the truncation index.}

\item{is.sorted}{Logical, indicating whether the supplied p-values are already
is.sorted. Defaults to FALSE.}

\item{earlyStop}{Logical; set to TRUE to stop as soon as a hypothesis can be
accepted at level alpha. This speeds up the procedure, but now only provides
lower bounds on the p-values for the global test.}

\item{verbose}{Logical; set to TRUE to print progress.}

\item{mc.cores}{Number of cores to parallelize onto. If mc.cores > 1 the procedure
cannot stop early and may then be slower than with mc.cores = 1.}

\item{...}{Additional arguments}
}
\value{
a data.frame containing adjusted p-values and their respective indices.
}
\description{
Adjust all marginally significant using a Closed Testing Procedeure and the
TMTI family of tests.
}
\examples{
p = sort(runif(100))  # Simulate and sort p-values
p[1:10] = p[1:10]**3  # Make the bottom 10 smaller, such that they correspond to false hypotheses
adjust_TMTI(p, alpha = 0.05, is.sorted = TRUE)
}
